Store decoded symbols in recoded file, first roundtrip!
Very hacky patch with a lot of #if/#else blocks to replace the
CABAC-encoded block in the recoded file with the raw symbols.
I had to implement bit-packing (which was difficult to debug) because
protobuf can't handle huge files.

This encoding demonstrates the technique working end-to-end. Now all
that remains is to encode the symbol stream efficiently.

+++ cabac_code.h
@@ -14,9 +14,9 @@ static const uint8_t * const ff_h264_mlps_state = ff_h264_cabac_tables + H264_ML
 
 
 struct cabac {
-  // Word size for encoder/decoder state.
-  typedef uint64_t fixed_point;
-  // Word size for compressed data.
+  // Word size for encoder/decoder state. Reasonable values: uint64_t, uint32_t.
+  typedef uint32_t fixed_point;
+  // Word size for compressed data. Reasonable values: uint16_t, uint8_t.
   typedef uint16_t compressed_digit;
   // min_range must be at least 0x200 so that range/2 never rounds in put_bypass.
   static constexpr int min_range = 0x200;
